% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{add_banner}
\alias{add_banner}
\title{Add Tracking Banner to Plot}
\usage{
add_banner(gg, tracker, height_plot = 7)
}
\arguments{
\item{gg}{ggplot object to track}

\item{tracker}{ggtrack tracker object}

\item{height_plot}{numeric tracker height in cm.}
}
\value{
tracker
}
\description{
Add Tracking Banner to Plot
}
\examples{

\dontrun{

  track <- make_tracker()

  ggplot() \%>\%
    add_banner(track)

}
}
