% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{add_banner_interactive}
\alias{add_banner_interactive}
\title{Add Banner to Interactive Chart}
\usage{
add_banner_interactive(
  gg,
  tracker,
  plot_height = 7,
  width_svg = 8,
  height_svg = 4,
  options
)
}
\arguments{
\item{gg}{ggplot object, interactivity needs to be added via elements such as \link[ggiraph]{geom_point_interactive}}

\item{tracker}{ggtrack tracker object}

\item{plot_height}{numeric height of the main plot panel in cm.}

\item{width_svg}{numeric svg width in inches, passed to \link[ggiraph]{girafe}}

\item{height_svg}{numeric svg width in inches, passed to \link[ggiraph]{girafe}}

\item{options}{list passed to \link[ggiraph]{girafe} options}
}
\value{
ggiraph svg object
}
\description{
Interactive and SVG plots are made using a combination
of \link\link{ggiraph} and \link\link{patchwork}. This is a very simple wrapper
for ..., in some instances it may be easier to use \link\link{ggiraff}
directly. See ggiraff documentation and examples at:
https://davidgohel.github.io/ggiraph/articles/offcran/examples.html
}
