% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caption.R
\name{add_caption}
\alias{add_caption}
\title{Add Caption to Tracker}
\usage{
add_caption(tracker, caption, position)
}
\arguments{
\item{caption}{\code{character} or \code{grob} to add to footer. Text can be
\code{html} or \code{md} and is passed directly to \link[gridtext]{richtext_grob}}

\item{position}{data.frame generated by get \link[ggtrack]{get_positions}}
}
\value{

}
\description{
Add Caption to Tracker
}
