% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{add_logo}
\alias{add_logo}
\title{Add Logo to Tracker}
\usage{
add_logo(tracker, logo, height_tracker, position, justification)
}
\arguments{
\item{position}{data.frame generated by get \link[ggtrack]{get_positions}}

\item{justification}{numeric between 0 and 1 passed to \link[grid]{rasterGrob}.}
}
\value{

}
\description{
Add Logo to Tracker
}
