% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{add_theme}
\alias{add_theme}
\title{Modify Tracking Banner Theme}
\usage{
add_theme(tracker, ...)
}
\arguments{
\item{tracker}{ggtrack tracker object}

\item{...}{options passed to \link[ggplot2]{theme} in order to style the tracker banner}
}
\value{
tracker
}
\description{
Modify Tracking Banner Theme
}
\examples{
\dontrun{
  make_tracker() \%>\% add_theme(plot.background = ggplot2::element_rect(fill = "red", size = 0))
}
}
