% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caption.R
\name{caption}
\alias{caption}
\alias{add_caption}
\alias{add_caption.gg}
\alias{add_caption.tracker}
\title{Add Caption to Tracker}
\usage{
caption(tracker, caption, position, ...)

add_caption(tracker, ...)

\method{add_caption}{gg}(tracker, caption, position, ...)

\method{add_caption}{tracker}(tracker, caption, ...)
}
\arguments{
\item{tracker}{ggtrack tracker object}

\item{caption}{\code{character} or \code{grob} to add to footer. Text can be
\code{html} or \code{md} and is passed directly to \link[gridtext]{richtext_grob}}

\item{position}{data.frame generated by get \link[ggtrack]{get_positions}}

\item{...}{additional options passed through to \link[gridtext]{richtext_grob}}
}
\value{
tracker
}
\description{
Add Caption to Tracker
}
\examples{
\dontrun{
  make_tracker() \%>\% add_caption('your caption')
}
}
\seealso{
Other gg: 
\code{\link{logo}()},
\code{\link{qr}()}

Other tracker: 
\code{\link{logo}()},
\code{\link{qr}()}

Other gg: 
\code{\link{logo}()},
\code{\link{qr}()}

Other tracker: 
\code{\link{logo}()},
\code{\link{qr}()}
}
\concept{add_caption}
\concept{gg}
\concept{tracker}
