% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_info.R
\name{get_git_info}
\alias{get_git_info}
\title{Get Local Git Information}
\usage{
get_git_info()
}
\value{
character
}
\description{
Uses \link[base]{system} to determine if a git repository exists
and if so returns the upstream remote URL and current commit SHA
}
\examples{
\dontrun{

get_git_info()

}
}
