% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\name{get_positions}
\alias{get_positions}
\title{Define Tracker Container Positions}
\usage{
get_positions(order, positions)
}
\arguments{
\item{order}{\code{character} of \code{length} 3 defining placement
order for "caption" (C), "logo" (L) and "QR" (Q) code. Must be one of:
\itemize{
\item{CLQ}
\item{LCQ}
\item{QLC}
\item{CQL}
}}

\item{positions}{\code{numeric} \code{vector} of \code{length} 3,
defining the horizontal proportion of each container. The 3 numbers
must add to 100.}
}
\value{

}
\description{
Define Tracker Container Positions
}
