% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{ggtrack}
\alias{ggtrack}
\title{Add Tracking Footer}
\usage{
ggtrack(
  gg,
  qr_content,
  color = "black",
  height_plot = 10,
  height_tracker = 1.8,
  interactive = FALSE,
  caption = NULL,
  logo = NULL,
  plotly_heights = c(0.8, 0.2)
)
}
\arguments{
\item{gg}{ggplot object to track}

\item{qr_content}{character to encode into QR code}

\item{color}{character color of the QR code}

\item{height_plot}{integer centimeters}

\item{height_tracker}{integer centimeters}

\item{interactive}{logical use plotly for interactivity}

\item{caption}{character text to add to footer}

\item{logo}{file to add to footer as a logo}

\item{plotly_heights}{numeric vector of length 2 to fix relative height of plot and tracking banner}
}
\value{

}
\description{
Add Tracking Footer
}
