% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr_hook.R
\name{ggtrack_hook}
\alias{ggtrack_hook}
\title{Tracker Knitr Hook}
\usage{
ggtrack_hook(x, options)
}
\arguments{
\item{x}{plot object}

\item{options}{knitr options}
}
\value{

}
\description{
Tracker Knitr Hook
}
