% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{make_download}
\alias{make_download}
\title{Tracker Download Link}
\usage{
make_download(
  tracker,
  save_file,
  download_file,
  ext = "png",
  link_text = "download",
  type = c("link", "button"),
  height = 13,
  height_units = c("cm", "mm", "in"),
  button_style = htmltools::css(background.color = "DodgerBlue", border = "none", color
    = "white", padding = "12px 30px", cursor = "pointer", font.size = "16px", width =
    "150px"),
  date = format(Sys.time(), "\%Y\%m\%d-\%H\%M\%S"),
  render = TRUE
)
}
\arguments{
\item{tracker}{tracker object to be saved}

\item{save_file}{character vector, where length > 1, path is generated using \link[base]{file.path}}

\item{download_file}{character vector, where length > 1, path is generated using \link[base]{file.path}}

\item{link_text}{character text to appear on the page}

\item{type}{character must be one of "link" or "button".}

\item{height}{integer height value passed to \link[ggplot2]{ggsave}}

\item{height_units}{character units value passed to \link[ggplot2]{ggsave}}

\item{button_style}{css object generated by \link[htmltools]{css} or a css character}

\item{render}{Boolean, render in-place or save for later}
}
\value{
html of the link object
}
\description{
Save tracker object to file and render link or button for user download
}
