% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr.R
\name{make_qr}
\alias{make_qr}
\title{Make QR}
\usage{
make_qr(qr_content, color = "black", color_bg = "white")
}
\arguments{
\item{qr_content}{\code{character} content passed to \link[qrencoder]{qrencode}. A time stamp is automatically added plus the current
git commit where available.}

\item{color_bg}{}
}
\value{

}
\description{
Make QR
}
