% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{plot.tracker}
\alias{plot.tracker}
\title{Plot Tracker Object}
\usage{
\method{plot}{tracker}(x, ...)
}
\arguments{
\item{x}{ggtrack tracker object}

\item{...}{plot options}
}
\description{
Plot Tracker Object
}
\examples{

plot(make_tracker())

}
