% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{print}
\alias{print}
\alias{print.tracker}
\title{Print Tracker}
\usage{
print(x, ...)

\method{print}{tracker}(tracker)
}
\description{
Print Tracker
}
\seealso{
Other tracker: 
\code{\link{caption}()},
\code{\link{logo}()},
\code{\link{qr}()}
}
\concept{print}
\concept{tracker}
