% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr.R
\name{qr}
\alias{qr}
\alias{add_qr}
\alias{add_qr.gg}
\alias{add_qr.tracker}
\title{Add QR Code to Tracker}
\usage{
qr(
  tracker,
  qr_content,
  color = "black",
  color_bg = "white",
  height_tracker,
  position,
  justification,
  ...
)

add_qr(tracker, ...)

\method{add_qr}{gg}(
  tracker,
  qr_content,
  color = "black",
  color_bg = "white",
  height_tracker,
  position,
  justification,
  ...
)

\method{add_qr}{tracker}(
  tracker,
  qr_content,
  color = "black",
  color_bg = "white",
  justification,
  ...
)
}
\arguments{
\item{tracker}{ggtrack tracker object}

\item{qr_content}{\code{character} content passed to \link[qrencoder]{qrencode}. A time stamp is automatically added plus the current
git commit where available.}

\item{color}{character color of QR code}

\item{color_bg}{character background color of QR code}

\item{height_tracker}{numeric tracker height in cm.}

\item{position}{data.frame generated by get \link[ggtrack]{get_positions}}

\item{justification}{numeric between 0 and 1 passed to \link[grid]{rasterGrob}.}

\item{...}{additional options passed to \link[grid]{rasterGrob}}
}
\value{
tracker
}
\description{
Add QR Code to Tracker
}
\examples{

  make_tracker() \%>\% add_qr('Report ID: 2c9075a5-4d7e-47a5-8616-55dd88af3dc5', justification = 1)

}
\seealso{
Other gg: 
\code{\link{caption}()},
\code{\link{logo}()}

Other tracker: 
\code{\link{caption}()},
\code{\link{logo}()}

Other gg: 
\code{\link{caption}()},
\code{\link{logo}()}

Other tracker: 
\code{\link{caption}()},
\code{\link{logo}()}
}
\concept{add_qr}
\concept{gg}
\concept{tracker}
