% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\docType{data}
\name{qr_mod}
\alias{qr_mod}
\title{Estimate QR Size.}
\format{
A linear model
}
\usage{
qr_mod
}
\description{
A model predicting the required QR size based
on the number of bytes to be encoded. The model is generated in the
package vignette qr_calibrate
}
\keyword{datasets}
