% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{qr_size}
\alias{qr_size}
\title{Get QR Size}
\usage{
qr_size(txt)
}
\arguments{
\item{txt}{character to be encoded into the QR}
}
\value{
numeric size of QR in cm.
}
\description{
Given a character string, calculate the minimum size
of a QR that will reliably scan off a computer monitor
}
