% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banner.R
\name{add_banner}
\alias{add_banner}
\title{Add Tracking Banner to Plot}
\usage{
add_banner(gg, tracker, height_plot = 7, banner_layout = NULL)
}
\arguments{
\item{gg}{ggplot object to track}

\item{tracker}{ggtrack tracker object}

\item{height_plot}{numeric tracker height in cm.}

\item{banner_layout}{plot_layout object created by \link[patchwork]{plot_layout}. Only
required where input is a \code{patchwork} object. Must include a banner height element
that is available from the \code{ggtrack} object via \verb{$height}. The banner is added to
the patchwork object via \code{/}.}
}
\value{
tracker
}
\description{
Add Tracking Banner to Plot
}
\examples{

\dontrun{

  track <- make_tracker()

  ggplot() \%>\%
    add_banner(track)

}
}
