% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{add_logo}
\alias{add_logo}
\title{Add Logo to Tracker}
\usage{
add_logo(tracker, logo, height_tracker, position)
}
\arguments{
\item{position}{data.frame generated by get \link[ggtrack]{get_positions}}
}
\value{

}
\description{
Add Logo to Tracker
}
