% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr.R
\name{add_qr}
\alias{add_qr}
\title{Add QR Code to Tracker}
\usage{
add_qr(tracker, qr_content, color, height_tracker, position, justification)
}
\arguments{
\item{qr_content}{\code{character} content passed to \link[qrencoder]{qrencode}. A time stamp is automatically added plus the current
git commit where available.}

\item{position}{data.frame generated by get \link[ggtrack]{get_positions}}

\item{justification}{numeric between 0 and 1 passed to \link[grid]{rasterGrob}.}
}
\value{

}
\description{
Add QR Code to Tracker
}
