% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{ggtrack}
\alias{ggtrack}
\title{Add Tracking Footer}
\usage{
ggtrack(
  gg,
  qr_content,
  color = "black",
  color_bg = "white",
  caption = NULL,
  logo = NULL,
  order = "CLQ",
  positions = c(55, 25, 20),
  logo_justification = 1,
  qr_justification = 1,
  height_plot = 7,
  height_tracker = 1.8,
  interactive = FALSE,
  plotly_heights = c(0.8, 0.2),
  ...
)
}
\arguments{
\item{gg}{ggplot object to track}

\item{qr_content}{\code{character} content passed to \link[qrencoder]{qrencode}. A time stamp is automatically added plus the current
git commit where available.}

\item{color}{character color of the QR code}

\item{color_bg}{character background color of QR code}

\item{caption}{\code{character} or \code{grob} to add to footer. Text can be
\code{html} or \code{md} and is passed directly to \link[gridtext]{richtext_grob}}

\item{logo}{file to add to footer as a logo}

\item{order}{\code{character} of \code{length} 3 defining placement
order for "caption" (C), "logo" (L) and "QR" (Q) code. Must be one of:
\itemize{
\item{CLQ}
\item{LCQ}
\item{QLC}
\item{CQL}
}}

\item{positions}{\code{numeric} \code{vector} of \code{length} 3,
defining the horizontal proportion of each container. The 3 numbers
must add to 100.}

\item{logo_justification}{numeric between 0 and 1 passed to \link[grid]{rasterGrob}. See note below.}

\item{qr_justification}{numeric between 0 and 1 passed to \link[grid]{rasterGrob}. See note below.}

\item{height_plot}{numeric plot height in cm.}

\item{height_tracker}{numeric tracker height in cm.}

\item{interactive}{logical use plotly for interactivity}

\item{plotly_heights}{numeric vector of length 2 to fix relative height of plot and tracking banner}

\item{...}{options passed to \link[ggplot2]{theme} in order to style the tracker banner.}
}
\value{
grid
}
\description{
Add Tracking Footer
}
\note{
For Justification you need to imagine each element in its own rectangle with a bottom
dimension of 0 to 1. If you want the logo or QR code on the right of the rectangle set justification
to 1, or 0 for left.
}
