% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{make_download}
\alias{make_download}
\title{Tracker Download Link}
\usage{
make_download(
  tracker,
  filename,
  ext = "png",
  link_text = "download",
  type = c("link", "button"),
  render = TRUE
)
}
\arguments{
\item{tracker}{tracker object to be saved}

\item{filename}{character vector, where length > 1, path is generated using \link[base]{file.path}}

\item{link_text}{character text to appear on the page}

\item{type}{character must be one of "link" or "button".}

\item{render}{Boolean, render in-place or save for later}
}
\value{
html of the link object
}
\description{
Save tracker object to file and render link or button for user download
}
