% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{make_tracker}
\alias{make_tracker}
\title{Define Tracker Base}
\usage{
make_tracker(order = "CLQ", positions = c(55, 25, 20), height_tracker = 1.8)
}
\arguments{
\item{height_tracker}{}
}
\value{
tracker
}
\description{
Define Tracker Base
}
