% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{obj_tracker}
\alias{obj_tracker}
\title{Create Tracker Object}
\usage{
obj_tracker(tracker, pos, height_tracker, contains)
}
\arguments{
\item{contains}{character vector tracks what elements have been added to tracker}
}
\value{

}
\description{
Create Tracker Object
}
