% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{print.tracker}
\alias{print.tracker}
\title{Print Tracker}
\usage{
\method{print}{tracker}(x, ...)
}
\arguments{
\item{x}{ggtrack tracker object}

\item{...}{print options}
}
\description{
Print Tracker
}
\examples{
\dontrun{
  print(make_tracker())
}
}
